requirements.txt
torch==1.9.0
torchvision==0.10.0
mmcv-full==1.3.12
prettytable